# packages/shared-types/openapi.json
{
  "components": {
    "schemas": {
...
      "DeviceInfo": {
        "additionalProperties": false,
        "properties": {
          "app_version": {
            "anyOf": [
              {
                "maxLength": 40,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "App Version"
          },
          "model": {
            "anyOf": [
              {
                "maxLength": 80,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "os": {
            "anyOf": [
              {
                "maxLength": 80,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Os"
          }
        },
        "title": "DeviceInfo",
        "type": "object"
      },
      "DeviceList": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/DeviceOut"
            },
            "title": "Data",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "data"
        ],
        "title": "DeviceList",
        "type": "object"
      },
      "DeviceOut": {
        "properties": {
          "app_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "App Version"
          },
          "connection": {
            "enum": [
              "online",
              "offline",
              "needs_pairing"
            ],
            "title": "Connection",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "guest",
              "kitchen"
            ],
            "title": "Kind",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "last_seen_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Seen At"
          },
          "os": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Os"
          },
          "paired_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Paired At"
          },
          "room": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RoomRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "stations": {
            "items": {
              "$ref": "#/components/schemas/Ref"
            },
            "title": "Stations",
            "type": "array"
          },
          "status": {
            "enum": [
              "active",
              "locked",
              "disabled",
              "reset_required",
              "revoked"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "label",
          "kind",
          "status",
          "connection",
          "room",
          "stations",
          "os",
          "app_version",
          "last_seen_at",
          "paired_at",
          "created_at"
        ],
        "title": "DeviceOut",
        "type": "object"
      },
      "DeviceTokenRequest": {
        "additionalProperties": false,
        "properties": {
          "device_credential": {
            "maxLength": 200,
            "minLength": 10,
            "title": "Device Credential",
            "type": "string"
          }
        },
        "required": [
          "device_credential"
        ],
        "title": "DeviceTokenRequest",
        "type": "object"
      },
      "DeviceTokenResponse": {
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "device_id": {
            "format": "uuid",
            "title": "Device Id",
            "type": "string"
          },
          "expires_in": {
            "title": "Expires In",
            "type": "integer"
          },
          "token_type": {
            "const": "Bearer",
            "default": "Bearer",
            "title": "Token Type",
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "expires_in",
          "device_id"
        ],
        "title": "DeviceTokenResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HeartbeatRequest": {
        "additionalProperties": false,
        "properties": {
          "app_version": {
            "anyOf": [
              {
                "maxLength": 40,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "App Version"
          },
          "battery": {
            "anyOf": [
              {
                "maximum": 100.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Battery"
          },
          "completed_commands": {
            "items": {
              "enum": [
                "RESET",
                "LOCK"
              ],
              "type": "string"
            },
            "maxItems": 4,
            "title": "Completed Commands",
            "type": "array"
          },
          "network": {
            "anyOf": [
              {
                "maxLength": 40,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Network"
          }
        },
        "title": "HeartbeatRequest",
        "type": "object"
      },
      "HeartbeatResponse": {
        "properties": {
          "commands": {
            "items": {
              "enum": [
                "RESET",
                "LOCK"
              ],
              "type": "string"
            },
            "title": "Commands",
            "type": "array"
          },
          "server_time": {
            "format": "date-time",
            "title": "Server Time",
            "type": "string"
          },
          "status": {
            "enum": [
              "active",
              "locked",
              "disabled",
              "reset_required",
              "revoked"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "commands",
          "server_time"
        ],
        "title": "HeartbeatResponse",
        "type": "object"
      },
...
      "PairRequest": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "pattern": "^\\d{6}$",
            "title": "Code",
            "type": "string"
          },
          "device_info": {
            "$ref": "#/components/schemas/DeviceInfo"
          }
        },
        "required": [
          "code"
        ],
        "title": "PairRequest",
        "type": "object"
      },
      "PairResponse": {
        "properties": {
          "device_credential": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Shown once. A replay of this request returns null; pair again with a new code.",
            "title": "Device Credential"
          },
          "device_id": {
            "format": "uuid",
            "title": "Device Id",
            "type": "string"
          },
          "hotel": {
            "$ref": "#/components/schemas/Ref"
          },
          "kind": {
            "enum": [
              "guest",
              "kitchen"
            ],
            "title": "Kind",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "room": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RoomRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "stations": {
            "items": {
              "$ref": "#/components/schemas/Ref"
            },
            "title": "Stations",
            "type": "array"
          }
        },
        "required": [
          "device_id",
          "device_credential",
          "label",
          "kind",
          "hotel",
          "room",
          "stations"
        ],
        "title": "PairResponse",
        "type": "object"
      },
      "PairingCreate": {
        "additionalProperties": false,
        "properties": {
          "room_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Room Id"
          },
          "station_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Station Id"
          },
          "station_ids": {
            "anyOf": [
              {
                "items": {
                  "format": "uuid",
                  "type": "string"
                },
                "maxItems": 20,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Station Ids"
          },
          "type": {
            "enum": [
              "guest",
              "kitchen"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type"
        ],
        "title": "PairingCreate",
        "type": "object"
      },
      "PairingOut": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "qr_payload": {
            "title": "Qr Payload",
            "type": "string"
          }
        },
        "required": [
          "code",
          "qr_payload",
          "expires_at"
        ],
        "title": "PairingOut",
        "type": "object"
      },
...
      "ReassignRequest": {
        "additionalProperties": false,
        "properties": {
          "room_id": {
            "format": "uuid",
            "title": "Room Id",
            "type": "string"
          }
        },
        "required": [
          "room_id"
        ],
        "title": "ReassignRequest",
        "type": "object"
      },
      "Ref": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "Ref",
        "type": "object"
      },
...
      "RoomRef": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "number": {
            "title": "Number",
            "type": "string"
          }
        },
        "required": [
          "id",
          "number"
        ],
        "title": "RoomRef",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Diyneco API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/devices": {
      "get": {
        "operationId": "list_devices_api_v1_devices_get",
        "parameters": [
          {
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "guest",
                    "kitchen"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "active",
                    "locked",
                    "disabled",
                    "reset_required",
                    "revoked"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "room_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Room Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 512,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Devices",
        "tags": [
          "devices"
        ]
      }
    },
    "/api/v1/devices/heartbeat": {
      "post": {
        "operationId": "heartbeat_api_v1_devices_heartbeat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HeartbeatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HeartbeatResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Heartbeat",
        "tags": [
          "devices"
        ]
      }
    },
    "/api/v1/devices/pair": {
      "post": {
        "description": "Redeems a pairing code. The credential is in this response only; a replay of the same\nrequest returns `device_credential: null` (DECISIONS D23).",
        "operationId": "pair_device_api_v1_devices_pair_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PairRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PairResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Pair Device",
        "tags": [
          "devices"
        ]
      }
    },
    "/api/v1/devices/pairings": {
      "post": {
        "operationId": "create_pairing_api_v1_devices_pairings_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PairingCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PairingOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Pairing",
        "tags": [
          "devices"
        ]
      }
    },
    "/api/v1/devices/token": {
      "post": {
        "description": "Exchanges the device credential for a 1-hour device token (security spec, Credentials).",
        "operationId": "device_token_api_v1_devices_token_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceTokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Device Token",
        "tags": [
          "devices"
        ]
      }
    },
    "/api/v1/devices/{device_id}": {
      "get": {
        "operationId": "get_device_api_v1_devices__device_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "device_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Device Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Device",
        "tags": [
          "devices"
        ]
      }
    },
    "/api/v1/devices/{device_id}/disable": {
      "post": {
        "operationId": "disable_device_api_v1_devices__device_id__disable_post",
        "parameters": [
          {
            "in": "path",
            "name": "device_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Device Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Disable Device",
        "tags": [
          "devices"
        ]
      }
    },
    "/api/v1/devices/{device_id}/lock": {
      "post": {
        "operationId": "lock_device_api_v1_devices__device_id__lock_post",
        "parameters": [
          {
            "in": "path",
            "name": "device_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Device Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Lock Device",
        "tags": [
          "devices"
        ]
      }
    },
    "/api/v1/devices/{device_id}/pairing": {
      "delete": {
        "operationId": "unpair_device_api_v1_devices__device_id__pairing_delete",
        "parameters": [
          {
            "in": "path",
            "name": "device_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Device Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Unpair Device",
        "tags": [
          "devices"
        ]
      }
    },
    "/api/v1/devices/{device_id}/reassign": {
      "post": {
        "operationId": "reassign_device_api_v1_devices__device_id__reassign_post",
        "parameters": [
          {
            "in": "path",
            "name": "device_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Device Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReassignRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reassign Device",
        "tags": [
          "devices"
        ]
      }
    },
    "/api/v1/devices/{device_id}/reset": {
      "post": {
        "operationId": "reset_device_api_v1_devices__device_id__reset_post",
        "parameters": [
          {
            "in": "path",
            "name": "device_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Device Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reset Device",
        "tags": [
          "devices"
        ]
      }
    },
    "/api/v1/devices/{device_id}/unlock": {
      "post": {
        "description": "Returns a locked or disabled device to active.",
        "operationId": "unlock_device_api_v1_devices__device_id__unlock_post",
        "parameters": [
          {
            "in": "path",
            "name": "device_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Device Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Unlock Device",
        "tags": [
          "devices"
        ]
      }
    },
...
  }
}